THEOPRASTUS VERSION 2.0
feat: Added InMemorySessionService to store user preferences in active session.

docs: Modified the README for documentation.
docs: Removed some script test used for diagnosis and debbugging.

refactor: Refactored subagent, agent, tools and utils scripts to fit the new features for the agent.
refactor: Merged tools and utils into other scripts and modules to make more manageable the development of the agent.

style: Made a few changes on the aspect of the script to be more readable.
style: Added comments for the description of some tools and utils.

chores: Moved and added folders for a better structure of the agent.